High order component for sidebar meta field
* Add a high level component function that can be used to connect an
input component with a certain post meta field.
* 2 examples of different usages: define the component separately or everything
in one place inside of render, each method with its pros and cons, but I
tend to prefer the convenience of having everything in one place.

This depends on changes in
https://github.com/greenpeace/planet4-master-theme , the 2 fields that
are included in this commit must be added using `register_post_meta()` and
must be removed from the metabox registered in `register_campaigns_metaboxes`,
else they are overwritten by the value at page load.

https://github.com/greenpeace/planet4-master-theme/tree/test-frontend-metabox