PLANET-4811 Only search DOM on frontend, useSelect in editor
* Since we only need to watch for changes in the editor, we don't need
to make the code that queries the DOM update the block when the DOM
changes. Instead we can use `useSelect` to get the blocks and filter the
headers (and potentially other blocks that can output a heading).
* Decouple fetching the headers from creating a hierarchical
representation.
* Just fetch the headings in render of SubmenuFrontend. This component
doesn't get re-rendered anyway, and also the headings won't change
(unless other blocks still need to be rendered, we probably should check
that).
* Split functions into dedicated files.
* Add backtop as an effect, which simply toggles the display of it. The
backtop markup and style will be added to the base template in
https://github.com/greenpeace/planet4-master-theme/pull/1166
* Use plain old links instead of implementing our own behavior. Mainly
did this to remove (a lot of) complexity from the code while doing other
changes.  We should probably evaluate if we still want to do a transition,
though this can probably still make use of anchors.